backend/requirements.txt
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
# AlgoTrade Pro — Python Dependencies
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

# ── Core Framework ──
fastapi==0.115.0
uvicorn[standard]==0.34.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.0

# ── Database (PostgreSQL) ──
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# ── Security & Auth ──
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography>=42.0.0
slowapi==0.1.9
python-multipart>=0.0.9

# ── HTTP Client ──
httpx==0.28.0

# ── Broker SDKs (Sprint 2) ──
smartapi-python>=1.5.5
kiteconnect>=5.0.1
pyotp>=2.9.0
logzero>=1.7.0
websocket-client>=1.6.0

# ── Data & Analysis ──
pandas>=2.2.0
numpy>=1.26.0

# ── Logging ──
python-json-logger==3.2.0
